refactored auth midlleware and imports
