+++ app/lib/di/splash_dependency.dart
@@ -1,14 +1,20 @@
 import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
-import 'package:matchpet/authentication_manager.dart';
+
+import '../services/authentication_manager.dart';
 
 class SplashDependency {
   Future<void> initializeSettings() async {
-    final AuthenticationManager _authManager = Get.put(AuthenticationManager());
+    //Inicializa o acesso ao storage
     await GetStorage.init();
+    //Injeta o Gerenciador de Autenticação
+    final AuthenticationManager _authManager = Get.put(AuthenticationManager());
+    //Verifica se o usuario ja estava logado
     _authManager.checkLoginStatus();
-    // await Future.delayed(const Duration(seconds: 2));
+    //Espera 3 segundos
+    await Future.delayed(const Duration(seconds: 3));
+    //Remove a Splash Screen
     FlutterNativeSplash.remove();
   }
 }
