+++ app/lib/pages/initial_page.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:get/get.dart';
-import 'package:matchpet/routes/app_routes.dart';
 import 'package:theme/export_theme.dart';
+
+import '../routes/app_routes.dart';
 
 class InitialPage extends StatelessWidget {
   const InitialPage({Key? key}) : super(key: key);
