--- app/lib/pages/splash_screen.dart
@@ -1,55 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-
-import '../authentication_manager.dart';
-import '../on_board.dart';
-
-class SplashScreen extends StatelessWidget {
-  final AuthenticationManager _authManager = Get.put(AuthenticationManager());
-
-  SplashScreen({Key? key}) : super(key: key);
-
-  Future<void> initializeSettings() async {
-    _authManager.checkLoginStatus();
-    await Future.delayed(const Duration(seconds: 1));
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder(
-      future: initializeSettings(),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.waiting) {
-          return waitingView();
-        } else {
-          if (snapshot.hasError) {
-            return errorView(snapshot);
-          } else {
-            return const OnBoard();
-          }
-        }
-      },
-    );
-  }
-
-  Scaffold errorView(AsyncSnapshot<Object?> snapshot) {
-    return Scaffold(body: Center(child: Text('Error: ${snapshot.error}')));
-  }
-
-  Scaffold waitingView() {
-    return Scaffold(
-        body: Center(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: const [
-          Padding(
-            padding: EdgeInsets.all(16.0),
-            child: CircularProgressIndicator(),
-          ),
-          Text('Loading...'),
-        ],
-      ),
-    ));
-  }
-}
