+++ app/lib/routes/app_pages.dart
@@ -1,12 +1,12 @@
-import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
-import 'package:matchpet/Pages/initial_page.dart';
-import 'package:matchpet/authentication_manager.dart';
-import 'package:matchpet/routes/app_routes.dart';
 import 'package:pet_profile/pages/pet_register_page.dart';
 import 'package:user_profile/pages/login_page.dart';
 import 'package:user_profile/pages/register_page.dart';
 import 'package:user_profile/pages/status_page.dart';
+
+import '../middlewares/auth_middleware.dart';
+import '../pages/initial_page.dart';
+import 'app_routes.dart';
 
 class AppPages {
   static final routes = [
@@ -32,14 +32,4 @@ class AppPages {
       page: () => PetRegisterPage(),
     ),
   ];
-}
-
-class AuthMiddleware extends GetMiddleware {
-  @override
-  RouteSettings? redirect(String? route) {
-    final AuthenticationManager _authManager = Get.find();
-    return _authManager.isLogged.value
-        ? const RouteSettings(name: Routes.statusRoute)
-        : null;
-  }
 }
