+++ app/lib/services/authentication_manager.dart
