+++ app/lib/services/cache_manager.dart
