+++ modules/theme/.flutter-plugins
@@ -1,3 +1,9 @@
 # This is a generated file; do not edit or check into version control.
 location=C:\\Users\\luis_\\AppData\\Local\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\location-4.4.0\\
 location_web=C:\\Users\\luis_\\AppData\\Local\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\location_web-3.1.1\\
+path_provider=C:\\Users\\luis_\\AppData\\Local\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider-2.0.11\\
+path_provider_android=C:\\Users\\luis_\\AppData\\Local\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_android-2.0.21\\
+path_provider_ios=C:\\Users\\luis_\\AppData\\Local\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_ios-2.0.11\\
+path_provider_linux=C:\\Users\\luis_\\AppData\\Local\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_linux-2.1.7\\
+path_provider_macos=C:\\Users\\luis_\\AppData\\Local\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_macos-2.0.6\\
+path_provider_windows=C:\\Users\\luis_\\AppData\\Local\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_windows-2.1.3\\
