+++ modules/user_profile/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"location","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\location-4.4.0\\\\","native_build":true,"dependencies":[]},{"name":"path_provider_ios","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_ios-2.0.11\\\\","native_build":true,"dependencies":[]}],"android":[{"name":"location","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\location-4.4.0\\\\","native_build":true,"dependencies":[]},{"name":"path_provider_android","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_android-2.0.21\\\\","native_build":true,"dependencies":[]}],"macos":[{"name":"location","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\location-4.4.0\\\\","native_build":true,"dependencies":[]},{"name":"path_provider_macos","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-2.0.6\\\\","native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-2.1.7\\\\","native_build":false,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-2.1.3\\\\","native_build":false,"dependencies":[]}],"web":[{"name":"location_web","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\location_web-3.1.1\\\\","dependencies":[]}]},"dependencyGraph":[{"name":"location","dependencies":["location_web"]},{"name":"location_web","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_ios","path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_ios","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2022-11-11 09:03:04.395744","version":"3.3.4"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"location","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\location-4.4.0\\\\","native_build":true,"dependencies":[]},{"name":"path_provider_ios","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_ios-2.0.11\\\\","native_build":true,"dependencies":[]}],"android":[{"name":"location","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\location-4.4.0\\\\","native_build":true,"dependencies":[]},{"name":"path_provider_android","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_android-2.0.21\\\\","native_build":true,"dependencies":[]}],"macos":[{"name":"location","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\location-4.4.0\\\\","native_build":true,"dependencies":[]},{"name":"path_provider_macos","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-2.0.6\\\\","native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-2.1.7\\\\","native_build":false,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-2.1.3\\\\","native_build":false,"dependencies":[]}],"web":[{"name":"location_web","path":"C:\\\\Users\\\\luis_\\\\AppData\\\\Local\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\location_web-3.1.1\\\\","dependencies":[]}]},"dependencyGraph":[{"name":"location","dependencies":["location_web"]},{"name":"location_web","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_ios","path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_ios","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2022-11-13 14:24:55.299207","version":"3.3.4"}
