+++ modules/user_profile/lib/pages/status_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:matchpet/routes/app_routes.dart';
-import 'package:matchpet/authentication_manager.dart';
-
+import 'package:matchpet/services/authentication_manager.dart';
 import 'package:theme/export_theme.dart';
-import 'package:user_profile/model/token.dart';
+
 import '../controller/user_controller.dart';
+import '../model/token.dart';
 import '../model/user.dart';
 
 class StatusPage extends StatelessWidget {
